Interface MultisetBuilder added Interface Builder improved

+++ src/main/java/io/github/deepeshpatel/jnumbertools/generator/combination/multiset/MultisetCombinationByRanks.java
@@ -28,7 +28,7 @@ import java.util.Map;
  * @param <T> the type of elements in the combinations; must be comparable for LEX/REVERSE_LEX ordering
  * @author Deepesh Patel
  */
-public class MultisetCombinationByRanks<T> extends AbstractMultisetCombination<T> {
+public class MultisetCombinationByRanks<T> extends AbstractMultisetCombination<T>  {
 
     /**
      * The iterable sequence of rank numbers to generate combinations.
@@ -41,7 +41,6 @@ public class MultisetCombinationByRanks<T> extends AbstractMultisetCombination<T
      * @param options a {@code LinkedHashMap} of distinct elements and their frequencies; insertion order affects INSERTION ordering
      * @param r the size of each combination (rᵣ); must be non-negative
      * @param ranks an iterable of 0-based rank numbers (0 ≤ rank < total combinations)
-     * @return a new MultisetCombinationByRanks instance
      * @throws IllegalArgumentException if rᵣ is negative, options is invalid, or ranks are out of bounds
      */
     public MultisetCombinationByRanks(LinkedHashMap<T, Integer> options, int r, Iterable<BigInteger> ranks) {
@@ -62,6 +61,7 @@ public class MultisetCombinationByRanks<T> extends AbstractMultisetCombination<T
     public Iterator<Map<T, Integer>> iterator() {
         return new SequenceIterator();
     }
+
 
     /**
      * Iterator for generating multiset combinations at specified ranks.
@@ -131,7 +131,7 @@ public class MultisetCombinationByRanks<T> extends AbstractMultisetCombination<T
                         remainingR -= x;
                         break;
                     } else {
-                        remainingRank -= ways;
+                        remainingRank -= (int) ways;
                     }
                 }
             }
